make EthereumNetworkRepository not importing android classes so that they can be used outside of android project (e.g. in dmz and util)
